+++ PrivateFrameworks/UpNextWidget.framework/UpNextWidget.VibrantLabel.h
@@ -16,7 +16,7 @@
 @property (nonatomic, retain) UIFont *font;
 @property (nonatomic, retain) _TtC12UpNextWidget25TintColorAsTextColorLabel *label;
 @property (nonatomic) int lineBreakMode;
-@property (nonatomic) long numberOfLines;
+@property (nonatomic) int numberOfLines;
 @property (nonatomic, copy) NSString *text;
 
 - (id /* block */).cxx_destruct;
@@ -26,12 +26,12 @@
 - (id)initWithFrame:(struct CGRect { struct CGPoint { float x_1_1_1; float x_1_1_2; } x1; struct CGSize { float x_2_1_1; float x_2_1_2; } x2; })arg1;
 - (id)label;
 - (int)lineBreakMode;
-- (long)numberOfLines;
+- (int)numberOfLines;
 - (void)setAttributedText:(id)arg1;
 - (void)setFont:(id)arg1;
 - (void)setLabel:(id)arg1;
 - (void)setLineBreakMode:(int)arg1;
-- (void)setNumberOfLines:(long)arg1;
+- (void)setNumberOfLines:(int)arg1;
 - (void)setText:(id)arg1;
 - (struct CGSize { float x1; float x2; })sizeThatFits:(struct CGSize { float x1; float x2; })arg1;
 - (id)text;
