+++ protocols/Swift._NSFastEnumeration.h
@@ -5,6 +5,6 @@
 
 @required
 
-- (long)countByEnumeratingWithState:(struct { unsigned long x1; id *x2; unsigned long x3; unsigned long x4[5]; }*)arg1 objects:(id*)arg2 count:(long)arg3;
+- (int)countByEnumeratingWithState:(struct { unsigned long x1; id *x2; unsigned long x3; unsigned long x4[5]; }*)arg1 objects:(id*)arg2 count:(int)arg3;
 
 @end
